cctv_scanner/requirements.txt
requests==2.31.0
dnspython==2.4.2
python-whois==0.8.0
scapy>=2.5.0
flask==3.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
python-nmap>=0.7.1
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
python-rtsp-client>=0.1.0
tkinter
pillow>=10.0.0 